requirements.txt
discord.py==2.2.3
python-dotenv==1.0.1
aiohttp~=3.9.1
openai==1.10.0